+++ src/Exception/LogicException.php
@@ -5,6 +5,6 @@ namespace Lmc\Matej\Exception;
 /**
  * Exception represents error in the program logic.
  */
-class LogicException extends AbstractMatejException
+class LogicException extends \LogicException implements MatejExceptionInterface
 {
 }