+++ src/Exception/MatejExceptionInterface.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace Lmc\Matej\Exception;
+
+/**
+ * Common interface for all exceptions thrown from Matej API Client.
+ */
+interface MatejExceptionInterface
+{
+}