+++ src/Exception/ResponseDecodingException.php
@@ -7,7 +7,7 @@ use Psr\Http\Message\ResponseInterface;
 /**
  * Exception thrown when Matej response cannot be decoded.
  */
-class ResponseDecodingException extends AbstractMatejException
+class ResponseDecodingException extends \RuntimeException implements MatejExceptionInterface
 {
     public static function forJsonError(string $jsonErrorMsg, ResponseInterface $response): self
     {
@@ -28,6 +28,36 @@ class ResponseDecodingException extends AbstractMatejException
             sprintf(
                 "Error decoding Matej response: required data missing.\n\nBody:\n%s",
                 $response->getBody()
+            )
+        );
+    }
+
+    public static function forInconsistentNumberOfCommands(int $numberOfCommands, int $commandResponsesCount): self
+    {
+        return new self(
+            sprintf(
+                'Provided numberOfCommands (%d) is inconsistent with actual count of command responses (%d)',
+                $numberOfCommands,
+                $commandResponsesCount
+            )
+        );
+    }
+
+    public static function forInconsistentNumbersOfCommandProperties(
+        int $numberOfCommands,
+        $numberOfSuccessfulCommands,
+        $numberOfFailedCommands,
+        $numberOfSkippedCommands
+    ): self {
+        return new self(
+            sprintf(
+                'Provided numberOfCommands (%d) is inconsistent with provided sum of '
+                . 'numberOfSuccessfulCommands (%d) + numberOfFailedCommands (%d)'
+                . ' + numberOfSkippedCommands (%d)',
+                $numberOfCommands,
+                $numberOfSuccessfulCommands,
+                $numberOfFailedCommands,
+                $numberOfSkippedCommands
             )
         );
     }