+++ tests/Model/CommandResponseTest.php
@@ -2,7 +2,7 @@
 
 namespace Lmc\Matej\Model;
 
-use Lmc\Matej\Exception\InvalidDomainModelArgumentException;
+use Lmc\Matej\Exception\ResponseDecodingException;
 use Lmc\Matej\TestCase;
 
 class CommandResponseTest extends TestCase
@@ -89,7 +89,7 @@ class CommandResponseTest extends TestCase
     /** @test */
     public function shouldThrowExceptionIfStatusIsMissing(): void
     {
-        $this->expectException(InvalidDomainModelArgumentException::class);
+        $this->expectException(ResponseDecodingException::class);
         $this->expectExceptionMessage('Status field is missing in command response object');
         CommandResponse::createFromRawCommandResponseObject((object) ['message' => 'Foo', 'data' => [['bar']]]);
     }